Refactor ApiService
- Rename ApiService to cypherQueryApiService
- Simplify the service
- Rename SchemaApiService to schemaApiService.ts
